Legend Generalizing and Legend Icon bug Fix
This build on my previous pull request patch.

This fixes a bug in the legend ( the shapes and colours were not
displaying ), and further removes a number of repeated code by making
radar.js more generic in nature.

* The h,w values have been moved into the radardData.js
* The hard coded 4 legends have been replaced with a single function,
which now supports N legends.
* Legend Positional information has been moved into RadarData.js
* The legends now correctly displays the icons and colours
* URL support has been added to the URL legends.

NOTE: I could not see a way to add URL click support to the labels since
they don't support events, but I am sure there must be a simple way
around this.